refactor: split app site manager helpers
